% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acroname.R
\name{acroname}
\alias{acroname}
\title{Create Names using Acronyms}
\usage{
acroname(required = list(), optional = list(), dict = dictionary())
}
\arguments{
\item{required}{list of required words. Words that are alternatives for each
other should be a character vector entry to the list.}

\item{optional}{list of optional words. Words that are alternatives for each
other should be a character vector entry to the list.}

\item{dict}{the dictionary of acceptable words. Must contain a column with
the name word.}
}
\value{
vector of possibilities
}
\description{
Create Names using Acronyms
}
\examples{
test <- acroname(required = list('Harvard', c('Automated', 'Algorithmic'),
'Redistricting', c('Team', 'Project', 'Group')), optional = list('Methodology'))
}
