% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Dictionary of words}
\format{
A data frame with 175393 rows and 2 variables:
\describe{
\item{word}{words from \code{words::words}}
\item{chars}{a letters list of characters}
}
}
\usage{
dict
}
\description{
A dataset words extracted and separated from \code{words::words}.
}
\keyword{datasets}
