% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Create a Dictionary}
\usage{
dictionary(dict = words::words)
}
\arguments{
\item{dict}{tibble with column \code{word}}
}
\value{
tibble
}
\description{
Create a Dictionary
}
\examples{
\donttest{
dict <- dictionary()
}
}
