% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letters.R
\name{letters_list}
\alias{letters_list}
\title{Create Letters List}
\usage{
letters_list(word)
}
\arguments{
\item{word}{list of words}
}
\value{
list
}
\description{
Create Letters List
}
\examples{
letters_list('abacus')
}
